Move the application logic out of `main.go`
